Fix a problem where expectations don't match reality.
I had to fix a problem where my test was expecting one method, "create", but the interface of the class had called it "initialize". This is the sort of error that can arise when using so many mocks like I am doing. It either suggests I need to use fewer mocks(would be nice) or find better ways to verify the beheviors classes expect are available.